requirements.txt
selenium>=4.0.0
requests>=2.0.0
python-dotenv>=0.19.0
